Further restructure of classes - FIX
