--- Luta.java
@@ -1,10 +1,0 @@
-package jogo.ufc;
-
-import java.util.ArrayList;
-
-public class Luta {
-
-    private ArrayList<Round> rounds;
-
-
-}
