--- Round.java
@@ -1,12 +1,0 @@
-package jogo.ufc;
-
-public class Round {
-
-    private Lutador lutador1;
-    private Lutador lutador2;
-
-    public void setLutador1(){
-
-    }
-
-}
