+++ jogo/ufc/Luta.java
@@ -1,4 +1,10 @@
 package jogo.ufc;
 
+import java.util.ArrayList;
+
 public class Luta {
+
+    private ArrayList<Round> rounds;
+
+
 }
