+++ jogo/ufc/Main.java
@@ -6,14 +6,14 @@ public class Main {
 
 	public static void main(String[] args) {
 		
-		Jogo test = new Jogo();
+		//Jogo test = new Jogo();
 		LeitorArquivo a = new LeitorArquivo();
 		try {
 			a.test();
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		test.start();
+		//test.start();
 	}
 
 }
