+++ jogo/ufc/Poderes.java
