+++ jogo/ufc/Round.java
@@ -1,4 +1,12 @@
 package jogo.ufc;
 
 public class Round {
+
+    private Lutador lutador1;
+    private Lutador lutador2;
+
+    public void setLutador1(){
+
+    }
+
 }
