+++ jogo/ufc/TodasAcoes.java
